refactor(infractions): Move infraction API to its own class
